requirements.txt
docopt==0.6.2
pygit2==1.7.0
PyYAML==5.4.1
